# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "博客系统的api接口文档",
        "title": "博客接口文档说明",
        "contact": {},
        "license": {},
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/articles": {
            "get": {
                "description": "获取文章列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取文章列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页数",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "标题",
                        "name": "title",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {}
                }
            },
            "post": {
                "description": "新增文章",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "新增文章",
                "parameters": [
                    {
                        "description": "新增文章",
                        "name": "state",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.AddArticleForm"
                        }
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/articles/{id}": {
            "get": {
                "description": "获取单个文章",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取单个文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            },
            "put": {
                "description": "修改文章",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "修改文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "唯一id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "新增文章",
                        "name": "state",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.EditArticleForm"
                        }
                    }
                ],
                "responses": {
                    "200": {}
                }
            },
            "delete": {
                "description": "删除文章",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "删除文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "唯一id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/signin": {
            "post": {
                "description": "登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "登录参数",
                        "name": "string",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.AuthForm"
                        }
                    }
                ],
                "responses": {
                    "200": {},
                    "500": {}
                }
            }
        },
        "/tags": {
            "get": {
                "description": "获取文章标签列表 带字段筛选 带分页",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取文章标签列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "当前页数",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "标签名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "标签名称",
                        "name": "state",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {}
                }
            },
            "post": {
                "description": "新增文章标签",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "新增文章标签",
                "parameters": [
                    {
                        "description": "新增文章标签",
                        "name": "state",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.AddTagForm"
                        }
                    }
                ],
                "responses": {
                    "200": {},
                    "500": {}
                }
            }
        },
        "/tags/{id}": {
            "put": {
                "description": "修改文章标签",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "修改文章标签",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "新增文章标签",
                        "name": "state",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.EditTagForm"
                        }
                    }
                ],
                "responses": {
                    "200": {}
                }
            },
            "delete": {
                "description": "删除文章标签",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "删除文章标签",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        }
    },
    "definitions": {
        "admin.AuthForm": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "v1.AddArticleForm": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "内容",
                    "type": "string"
                },
                "cover_image_url": {
                    "description": "封面图片",
                    "type": "string"
                },
                "created_by": {
                    "description": "创建人",
                    "type": "string"
                },
                "desc": {
                    "description": "描述",
                    "type": "string"
                },
                "state": {
                    "description": "启用or 禁用",
                    "type": "integer"
                },
                "tag_id": {
                    "description": "分类id",
                    "type": "integer"
                },
                "title": {
                    "description": "标题",
                    "type": "string"
                }
            }
        },
        "v1.AddTagForm": {
            "type": "object",
            "properties": {
                "created_by": {
                    "description": "创建人",
                    "type": "string"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "state": {
                    "description": "禁用 or 启用",
                    "type": "integer"
                }
            }
        },
        "v1.EditArticleForm": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "内容",
                    "type": "string"
                },
                "cover_image_url": {
                    "description": "封面图片",
                    "type": "string"
                },
                "desc": {
                    "description": "描述",
                    "type": "string"
                },
                "id": {
                    "description": "id",
                    "type": "integer"
                },
                "modified_by": {
                    "description": "修改人",
                    "type": "string"
                },
                "state": {
                    "description": "启用or 禁用",
                    "type": "integer"
                },
                "tag_id": {
                    "description": "分类id",
                    "type": "integer"
                },
                "title": {
                    "description": "标题",
                    "type": "string"
                }
            }
        },
        "v1.EditTagForm": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer"
                },
                "modified_by": {
                    "description": "修改人",
                    "type": "string"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "state": {
                    "description": "禁用 or 启用",
                    "type": "integer"
                }
            }
        }
    }
}